+++ src/mesido/workflows/goals/minimize_tco_goal.py
@@ -108,6 +108,10 @@ class MinimizeTCO(Goal):
         If `discounted_annualized_cost` is not defined or False, the operational and
         fixed operational costs are multiplied by the number_of_years.
 
+        Default value of 'technicalLifetime' is 0. If the default value is not modified
+        in ESDL file, 'technicalLifetime' is taken as 'number_of_years' of optimization
+        in the cost calculation
+
         Args:
             optimization_problem (TechnoEconomicMixin): The optimization problem instance.
             cost_type (str): The type of cost to calculate ("operational",
@@ -122,6 +126,11 @@ class MinimizeTCO(Goal):
         obj = 0.0
         for asset_type in asset_types:
             for asset in optimization_problem.energy_system_components.get(asset_type, []):
+                asset_technicalLifetime = optimization_problem.esdl_assets.get(
+                    optimization_problem.esdl_asset_name_to_id_map[asset], []
+                ).attributes["technicalLifetime"]
+                if asset_technicalLifetime == 0:
+                    asset_technicalLifetime = self.number_of_years
                 extra_var = optimization_problem.extra_variable(cost_type_map[asset])
                 if options["discounted_annualized_cost"]:
                     # We only want the operational cost for a single year when we use
@@ -131,7 +140,7 @@ class MinimizeTCO(Goal):
                     obj += extra_var * self.number_of_years
                 else:
                     # These are the CAPEX cost under non-annualized condition
-                    obj += extra_var
+                    obj += extra_var * (self.number_of_years / asset_technicalLifetime)
         return obj
 
     def function(self, optimization_problem: TechnoEconomicMixin, ensemble_member) -> MX: