+++ tests/test_end_scenario_sizing.py
@@ -52,6 +52,9 @@ class TestEndScenarioSizing(TestCase):
         - That buffer tank is only used on peak day
         - Check if TCO goal included the desired cost components.
 
+        - Check if lifetime of an asset is taken into account in
+        optimization for CAPEX and Fixed OPEX
+
 
         Missing:
         - Link ATES t0 utilization to state of charge at end of year for optimizations over one
@@ -100,12 +103,25 @@ class TestEndScenarioSizing(TestCase):
             *self.solution.energy_system_components.get("heat_pump", []),
             *self.solution.energy_system_components.get("heat_pipe", []),
         ]:
-            obj += self.results[f"{self.solution._asset_fixed_operational_cost_map[asset]}"] * years
+            technicalLifetime = self.solution.esdl_assets.get(
+                self.solution.esdl_asset_name_to_id_map[asset], []
+            ).attributes["technicalLifetime"]
+            if technicalLifetime == 0:
+                technicalLifetime = years
+            obj += (
+                self.results[f"{self.solution._asset_fixed_operational_cost_map[asset]}"]
+                * years
+                * (years / technicalLifetime)
+            )
             obj += (
                 self.results[f"{self.solution._asset_variable_operational_cost_map[asset]}"] * years
             )
-            obj += self.results[f"{self.solution._asset_investment_cost_map[asset]}"]
-            obj += self.results[f"{self.solution._asset_installation_cost_map[asset]}"]
+            obj += self.results[f"{self.solution._asset_investment_cost_map[asset]}"] * (
+                years / technicalLifetime
+            )
+            obj += self.results[f"{self.solution._asset_installation_cost_map[asset]}"] * (
+                years / technicalLifetime
+            )
 
         np.testing.assert_allclose(obj / 1.0e6, self.solution.objective_value)
 
@@ -128,6 +144,9 @@ class TestEndScenarioSizing(TestCase):
         - Unstaged approaches, using the general function run_optimization_problem and the
         function run_end_scenario_sizing with staged_pipe_optimization to False should have
         comparable computation times.
+
+        - Check if lifetime of an asset is taken into account in optimization for
+         CAPEX and Fixed OPEX
 
         Missing:
         - Link ATES t0 utilization to state of charge at end of year for optimizations over one
@@ -190,10 +209,23 @@ class TestEndScenarioSizing(TestCase):
             *solution_staged.energy_system_components.get("heat_pump", []),
             *solution_staged.energy_system_components.get("heat_pipe", []),
         ]:
-            obj += results[f"{solution_staged._asset_fixed_operational_cost_map[asset]}"] * years
+            technicalLifetime = solution_staged.esdl_assets.get(
+                solution_staged.esdl_asset_name_to_id_map[asset], []
+            ).attributes["technicalLifetime"]
+            if technicalLifetime == 0:
+                technicalLifetime = years
+            obj += (
+                results[f"{solution_staged._asset_fixed_operational_cost_map[asset]}"]
+                * years
+                * (years / technicalLifetime)
+            )
             obj += results[f"{solution_staged._asset_variable_operational_cost_map[asset]}"] * years
-            obj += results[f"{solution_staged._asset_investment_cost_map[asset]}"]
-            obj += results[f"{solution_staged._asset_installation_cost_map[asset]}"]
+            obj += results[f"{solution_staged._asset_investment_cost_map[asset]}"] * (
+                years / technicalLifetime
+            )
+            obj += results[f"{solution_staged._asset_installation_cost_map[asset]}"] * (
+                years / technicalLifetime
+            )
 
         np.testing.assert_allclose(obj / 1.0e6, solution_staged.objective_value)
 
@@ -352,8 +384,8 @@ if __name__ == "__main__":
     start_time = time.time()
     a = TestEndScenarioSizing()
     a.setUpClass()
-    a.test_end_scenario_sizing()
+    # a.test_end_scenario_sizing()
     a.test_end_scenario_sizing_staged()
-    a.test_end_scenario_sizing_discounted()
-    a.test_end_scenario_sizing_head_loss()
+    # a.test_end_scenario_sizing_discounted()
+    # a.test_end_scenario_sizing_head_loss()
     print("Execution time: " + time.strftime("%M:%S", time.gmtime(time.time() - start_time)))